Only check for approximate result for gdk-pixbuf decoding
The results given by the gdk-pixbuf implementation seems to be flaky,
probably due to their implementation not being stable between versions
or not deterministic. Instead, require only that the images "look
similar". I have coded the "similarity" to be based on the average error
on the pixel values and a peak error allowed, compared to the same image
being encoded by blurhash-python (an official implementation of the
blurhash algorithm). For that, I have also written a small script
running blurhash-python on all `png` images in `data`.